moved the challenge coded to prevent a cicular dependancy.
