+++ frontend/src/app/App.tsx
@@ -9,7 +9,7 @@ import useToastWatcher from 'features/system/hooks/useToastWatcher';
 
 import FloatingOptionsPanelButtons from 'features/tabs/components/FloatingOptionsPanelButtons';
 import FloatingGalleryButton from 'features/tabs/components/FloatingGalleryButton';
-import { useChallengeWatcher } from '../features/system/store/challenge';
+import useChallengeWatcher from 'features/system/store/useChallengeWatcher';
 
 keepGUIAlive();
